Decouple depth from ray bouncing
This is the right thing to do : If bounce at all depth < n, unless n==1 then this is taking forever because we take into account indirect illumination to compute indirect illumination and so on and so forth. But n==1 prevents indirect illumination to be applied to things seen in mirrors or through glass. Decoupling like that prevents using IL to compute IL but still allows IL to be computed for transmitted/reflected rays.